Move the toolbar location to the top left corner of the GCode Viewer
