+++ web/components/widgets/gcode-viewer.styl
@@ -7,6 +7,13 @@
         left: 0;
         right: 0;
 
+        .btn-toolbar {
+            position: absolute;
+            top: 10px;
+            left: 10px;
+            z-index: 102;
+        }
+
         .joystick {
             position: absolute;
             top: 5px;
