+++ web/components/widgets/gcode.styl
@@ -2,10 +2,6 @@
     .widget-content {
         padding: 10px;
 
-        .btn-toolbar {
-            margin-bottom: 5px;
-        }
-
         .gcode-table {
             .public_fixedDataTable_topShadow {
                 background: none;
